+++ src/types/index.ts
@@ -1,5 +1,6 @@
 export type * from './CustomMaskOptions';
 export type * from './MaskConfig';
+export type * from './MaskResult';
 export type * from './MoneyMaskOptions';
 export type * from './CreditCardMaskOptions';
 export type * from './DatetimeMaskOptions';